separate ResourceChart into its own component
