+++ src/js/pages/nodes/NodeDetailPage.js
@@ -1,4 +1,3 @@
-import _ from 'underscore';
 import {Link} from 'react-router';
 import mixin from 'reactjs-mixin';
 /* eslint-disable no-unused-vars */
@@ -7,10 +6,7 @@ import React from 'react';
 import {RouteHandler} from 'react-router';
 import {StoreMixin} from 'mesosphere-shared-reactjs';
 
-import BarChart from '../../components/charts/BarChart';
-import Chart from '../../components/charts/Chart';
 import CompositeState from '../../structs/CompositeState';
-import Config from '../../config/Config';
 import DateUtil from '../../utils/DateUtil';
 import DescriptionList from '../../components/DescriptionList';
 import HealthTab from '../../components/HealthTab';
@@ -20,15 +16,11 @@ import MesosSummaryStore from '../../stores/MesosSummaryStore';
 import MesosStateStore from '../../stores/MesosStateStore';
 import NodeHealthStore from '../../stores/NodeHealthStore';
 import Page from '../../components/Page';
-import ResourceTypes from '../../constants/ResourceTypes';
 import SidePanels from '../../components/SidePanels';
 import StringUtil from '../../utils/StringUtil';
+import ResourceChart from '../../components/charts/ResourceChart';
 import TabsMixin from '../../mixins/TabsMixin';
 import TaskView from '../../components/TaskView';
-import Units from '../../utils/Units';
-
-// number to fit design of width vs. height ratio
-const WIDTH_HEIGHT_RATIO = 4.5;
 
 class NodeDetailPage extends mixin(InternalStorageMixin, TabsMixin, StoreMixin) {
   constructor() {
@@ -103,21 +95,19 @@ class NodeDetailPage extends mixin(InternalStorageMixin, TabsMixin, StoreMixin) 
     let states = MesosSummaryStore.get('states');
     let resources = states[`getResourceStatesFor${itemType}IDs`]([item.id]);
 
-    let charts = [
-      this.getResourceChart('cpus', resources),
-      this.getResourceChart('mem', resources),
-      this.getResourceChart('disk', resources)
-    ];
-
-    return charts.map(function (chart, i) {
-      return (
-        <div key={i} className="column-12 column-mini-4">
-          <div className="row chart">
-            {chart}
-          </div>
-        </div>
-      );
-    });
+    return (
+      <div className="row">
+        <ResourceChart
+          resourceName="cpus"
+          resources={resources} />
+        <ResourceChart
+          resourceName="mem"
+          resources={resources} />
+        <ResourceChart
+          resourceName="disk"
+          resources={resources} />
+      </div>
+    );
   }
 
   getLoadingScreen() {
@@ -148,78 +138,6 @@ class NodeDetailPage extends mixin(InternalStorageMixin, TabsMixin, StoreMixin) 
         <p className="flush">
           {`Did not find a node by the id "${this.props.params.nodeID}"`}
         </p>
-      </div>
-    );
-  }
-
-  getResourceChart(resource, totalResources) {
-    let colorIndex = ResourceTypes[resource].colorIndex;
-    let resourceLabel = ResourceTypes[resource].label;
-    let resourceData = [{
-      name: 'Alloc',
-      colorIndex: colorIndex,
-      values: totalResources[resource]
-    }];
-    let resourceValue = Units.formatResource(
-      resource, _.last(totalResources[resource]).value
-    );
-    let resourceIconClasses = `icon icon-sprite icon-sprite-medium
-      icon-sprite-medium-color icon-resources-${resourceLabel.toLowerCase()}`;
-
-    let axisConfiguration = {
-      x: {hideMatch: /^0$/},
-      y: {showPercentage: false, suffix: '%'}
-    };
-
-    let maxY = 5;
-    totalResources[resource].forEach(function (resourceTotal) {
-      if (resourceTotal.percentage > maxY) {
-        maxY = resourceTotal.percentage;
-      }
-    });
-
-    maxY *= 1.5; // Multiply by 150%
-    maxY /= 10; // Divide so that we can round to nearest tenth
-    maxY = Math.round(maxY); // Round
-    maxY *= 10; // Multiply so that we get a percentage number between 0-100
-    maxY = Math.min(100, maxY); // Don't let it be greater than 100%
-
-    return (
-      <div key={resource} className="column-12
-        flex-box-align-vertical-center
-        container-pod
-        container-pod-super-short
-        flush-top">
-        <div className="media-object-spacing-wrapper media-object-offset">
-          <div className="media-object media-object-align-middle">
-            <div className="media-object-item">
-              <i className={resourceIconClasses}></i>
-            </div>
-            <div className="media-object-item">
-              <h4 className="flush-top flush-bottom inverse">
-                {resourceValue}
-              </h4>
-              <span className={`side-panel-resource-label
-                  text-color-${colorIndex}`}>
-                {resourceLabel.toUpperCase()}
-              </span>
-            </div>
-          </div>
-        </div>
-
-        <Chart calcHeight={function (w) { return w / WIDTH_HEIGHT_RATIO; }}
-          delayRender={true}>
-          <BarChart
-            axisConfiguration={axisConfiguration}
-            data={resourceData}
-            inverseStyle={true}
-            margin={{top: 0, left: 43, right: 10, bottom: 40}}
-            maxY={maxY}
-            refreshRate={Config.getRefreshRate()}
-            ticksY={3}
-            xGridLines={0}
-            y="percentage" />
-        </Chart>
       </div>
     );
   }
@@ -324,9 +242,7 @@ class NodeDetailPage extends mixin(InternalStorageMixin, TabsMixin, StoreMixin) 
           <div className="container container-fluid container-pod container-pod-short container-pod-super-short-top flush">
             {this.getBasicInfo(node)}
             <div className="container-pod container-pod-short">
-              <div className="row">
-                {this.getCharts('Node', node)}
-              </div>
+              {this.getCharts('Node', node)}
             </div>
             <div className="container-pod container-pod-divider-bottom container-pod-divider-inverse container-pod-divider-bottom-align-right flush-top flush-bottom">
               <ul className="tabs list-inline flush-bottom inverse">
